+++ src/scenes/About/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import TeamSectionItem from './Team Section Item';
+import TeamMember from '../../components/Team Member';
 
 class About extends React.Component {
   constructor() {
@@ -13,10 +13,13 @@ class About extends React.Component {
     const { teamMembers } = this.state;
 
     const teamSectionItems = teamMembers.map(member => {
-      return <TeamSectionItem key={member} name={member} />;
+      return <div key={member} className="col-8 col-sm-5 text-center">
+          <TeamMember name={member} />
+        </div>;
     });
 
-    return <div className="bg-gray">
+    return (
+      <div className="bg-gray">
         <div className="container">
           <div className="row justify-content-center py-3 p-md-5">
             <div className="col-12 col-md-4 pb-3 pb-md-0 v-center">
@@ -26,25 +29,25 @@ class About extends React.Component {
             <div className="col-12 col-md-8">
               <p>
                 Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                Facilis, exercitationem voluptatem corrupti quo, accusamus,
-                est molestiae dolorum aliquam minus odio dignissimos labore
-                dicta iste aut! Eum harum deserunt voluptas blanditiis.
+                Facilis, exercitationem voluptatem corrupti quo, accusamus, est
+                molestiae dolorum aliquam minus odio dignissimos labore dicta
+                iste aut! Eum harum deserunt voluptas blanditiis.
               </p>
               <p>
                 Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                Facilis, exercitationem voluptatem corrupti quo, accusamus,
-                est molestiae dolorum aliquam minus odio dignissimos labore
-                dicta iste aut! Eum harum deserunt voluptas blanditiis.
+                Facilis, exercitationem voluptatem corrupti quo, accusamus, est
+                molestiae dolorum aliquam minus odio dignissimos labore dicta
+                iste aut! Eum harum deserunt voluptas blanditiis.
               </p>
               <p>
                 Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                Facilis, exercitationem voluptatem corrupti quo, accusamus,
-                est molestiae dolorum aliquam minus odio dignissimos labore
-                dicta iste aut! Eum harum deserunt voluptas blanditiis.
+                Facilis, exercitationem voluptatem corrupti quo, accusamus, est
+                molestiae dolorum aliquam minus odio dignissimos labore dicta
+                iste aut! Eum harum deserunt voluptas blanditiis.
               </p>
             </div>
           </div>
-        <div className="row justify-content-center py-3 p-md-5">
+          <div className="row justify-content-center py-3 p-md-5">
             <div className="col-12 col-md-4 pb-3 pb-md-0 v-center">
               <h1>Meet the Team</h1>
               <h5>Let's get acquainted</h5>
@@ -54,7 +57,8 @@ class About extends React.Component {
             </div>
           </div>
         </div>
-      </div>;
+      </div>
+    );
   }
 }