+++ changelog.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.0.2
 
+#### 2018-10-30
+
+### Added
+- Team member route
+- Team member profile page
+- Global team member component
+  - Team member component layout
+
+### Removed
+- About page team section item
+
 #### 2018-10-25
 
 ### Added